requirements.txt
fastapi
pyjwt
uvicorn
passlib
bcrypt==4.0.1